requirements.txt
pytest==8.3.4
selenium==4.28.1
requests==2.32.3
allure-pytest==2.13.5
allure-python-commons==2.13.5
webdriver-manager==4.0.2